extract span module
